Move availableActions to Organization
