+++ src/api/organization.js
@@ -1,6 +1,12 @@
 import gql from "graphql-tag";
 
 export const schema = gql`
+  type Action {
+    name: String
+    displayName: String
+    instructions: String
+  }
+
   type Organization {
     id: ID
     uuid: String
@@ -12,6 +18,7 @@ export const schema = gql`
     ): CampaignsReturn
     people(role: String, campaignId: String, sortBy: SortPeopleBy): [User]
     optOuts: [OptOut]
+    availableActions: [Action]
     threeClickEnabled: Boolean
     optOutMessage: String
     textingHoursEnforced: Boolean
